requirements.txt
anyio==4.9.0
certifi==2025.6.15
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dotenv==1.1.1
python-telegram-bot==22.2
sniffio==1.3.1
typing-extensions==4.14.0
